### 2024-02-17 #6363 @editorialbot
[PRE REVIEW]: pivmet: an `R` package proposing pivotal methods for consensus clustering and mixture modelling
**Submitting author:** @LeoEgidi (<a href="http://orcid.org/0000-0003-3211-905X">LEONARDO EGIDI</a>)
**Repository:** https://github.com/LeoEgidi/pivmet
**Branch with paper.md** (empty if default branch): master
**Version:** v0.5.0
