### 2024-03-10 #6363 @adriancorrendo
> Hi Adrian @adriancorrendo - will you be happy to review this submission? PDF can be found here [#6363 (comment)](https://github.com/openjournals/joss-reviews/issues/6363#issuecomment-1950276623).

Hi, @skanwal . Thanks for the invitation.
I'm interested in reviewing this.
Could you please remind me a reasonable timeline to complete the review?

Best regards,
Adrian
