### 2024-03-10 #6461 @editorialbot
[REVIEW]: pivmet: an R package proposing pivotal methods for consensus clustering and mixture modelling
**Submitting author:** @LeoEgidi (<a href="http://orcid.org/0000-0003-3211-905X">LEONARDO EGIDI</a>)
**Repository:** https://github.com/LeoEgidi/pivmet
**Branch with paper.md** (empty if default branch): master
**Version:** v0.6.0
