### 2024-03-24 #6461 @arfon
@LeoEgidi – this may have been discussed already but what license is this software released under? JOSS requires software published here to have a file named `LICENSE` at the root of the repository with the whole text of an OSI-approved license.
