### 2024-03-26 #6461 @LeoEgidi
Ok, now I put a LICENSE file at the repository root as suggested

> > @LeoEgidi – this may have been discussed already but what license is this software released under? JOSS requires software published here to have a file named `LICENSE` at the root of the repository with the whole text of an OSI-approved license.
>
> Hi @arfon , the License, as documented by the description file, is GPL-2
