+++ LICENSE.txt
@@ -1,0 +1,24 @@
+pivmet: R package for clustering and supervised learning
+
+Copyright (C) Leonardo Egidi
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU General Public License
+as published by the Free Software Foundation; either version 2
+of the License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program; if not, write to the Free Software
+Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+For any doubt or clarification about the use of the software, please 
+contact me, the author, at the following mail: legidi@units.it.
+
+Leonardo Egidi
+Author and mantainer
+Via Valerio 4/1 D, 34127 Trieste, Italy
