### 2024-04-05 LeoEgidi/pivmet#1 @adriancorrendo
JOSS reviewer comments - submission 6561
> ## Review checklist for @adriancorrendo

> ### General checks

> * [x]  **Contribution and authorship:** Has the submitting author (@LeoEgidi) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- Commits come only from @LeoEgidi. Could you please confirm the role of the rest of the authors?

> ### Functionality

1 > * [ ]  **Functionality:** Have the functional claims of the software been confirmed?

- I’m afraid there is no example in the paper of at least a line of code running any package functions. Could you please try to include something short with a line of code to run the key/s function/s?

2 > * [ ]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

- It sounds silly, but authors only indicate how to do it from github (although the code line to do it from CRAN is there). I think it could be good to use something like this on the README
:

** You can install the CRAN version of pivmet with:
[install.packages](https://rdrr.io/r/utils/install.packages.html)("pivmet")

** You can install the development version from [GitHub](https://github.com/) with: …

3 > * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

Pivmet repo does not include any automated tests through continuous integration tools. @LeoEgidi, could you please revise this point? I could not find these tests like: R-CMD, codecov, appveyor.

4 > * [ ]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I could not find community guidelines documentation. Please, see an example here…https://www.tidymodels.org/contribute/

> ### Software paper

5 > * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

I consider authors fail to provide a comparison of what potentially similar packages are available or mentioning how do people is figuring out for similar tasks without pivmet. Could you please offer a brief overview to readers?

6 > * [x]  **Typo on L59**

Please, could you check for a potential typo at line 59 where it says..."Figure ??"
