### 2024-04-05 #6461 @adriancorrendo
Hi @LeoEgidi .
How are you? Thanks for developing the pivmet package and submitting it for review here.

I've just created an issue on your repo with my review comments at https://github.com/LeoEgidi/pivmet/issues/1#issue-2226985183

Could you please take a look?

Thanks,
Adrian
