### 2024-04-08 #6461 @LeoEgidi
Hi @adriancorrendo , thanks a lot for your comments, I took a first look. I'll try to get a deeper look this week, or at most next week. Should I reply your points here or directly in the repo? What do you prefer? Thanks a lot
