### 2024-04-16 #6461 @larryshamalama
Hi @LeoEgidi, thanks for the submission. Overall, looks good, but I do have some doubts with the paper. Some comments below align with @adriancorrendo's review.

## Major comments/questions

- **A statement of need** and **State of the field**: You define _pivotal units_ in the first paragraph without diving into the details of the math; this is nice. Was this concept introduced in your 2018 paper or was it previously introduced? The main comment here is that I cannot find any "literature (software) review" of any packages that have any capabilities in finding pivotal units. If the concept was introduced in your 2018 paper, it makes sense that there is no software that specifically provides functionality for identifying pivotal units. It would be nice to either 1) if no other software can be used, identify what mixture modelling packages could potentially be helpful in identifying pivotal units or 2) summarize what this package contributes on top of other pivotal unit software.

## Minor comments

- I think that it would be good to include a quickstart akin to what is available in the README in the software paper itself.
- Community guidelines seem to be missing. For instance, you can add a `CONTRIBUTING.md` file outlining how to ask questions (via GitHub issues or discussions) and how to contribute to the package via pull requests.
- I found some typos in the article. I tried solving the ones I found in [my PR](https://github.com/LeoEgidi/pivmet/pull/2). I may have missed some, please have a look again.
