+++ README.md
@@ -1,50 +1,48 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
+```{r, echo = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-"
+)
+```
+
 # pivmet
 
-The goal of `pivmet` is to propose some pivotal methods in order to:
+The goal of ```pivmet``` is to propose some pivotal methods in order to:
 
-  - undo the label switching problem which naturally arises during the
-    MCMC sampling in Bayesian mixture models \(\rightarrow\) **pivotal
-    relabelling** (Egidi et al. 2018a)
+- undo the label switching problem which naturally arises during the MCMC sampling in Bayesian mixture models $\rightarrow$ **pivotal relabelling** (Egidi et al. 2018a)
 
-  - initialize the K-means algorithm aimed at obtaining a good
-    clustering solution \(\rightarrow\) **pivotal seeding** (Egidi et
-    al. 2018b)
+- initialize the K-means algorithm aimed at obtaining a good clustering solution $\rightarrow$ **pivotal seeding** (Egidi et al. 2018b)
 
 ## Installation
 
-  - <span style="color:red">PAY ATTENTION\! BEFORE INSTALLING</span>:
-    make sure to download the JAGS program at
-    <https://sourceforge.net/projects/mcmc-jags/>.
+- <span style="color:red">PAY ATTENTION! BEFORE INSTALLING</span>: make sure to download the JAGS program at
+[https://sourceforge.net/projects/mcmc-jags/](https://sourceforge.net/projects/mcmc-jags/).
 
-You can then install `pivmet` from github with:
+You can install the CRAN version of ```pivmet``` with:
 
-``` r
+```{r, eval = FALSE}
+install.packages("pivmet")
+library(pivmet)
+```
+
+You can install the development version of ```pivmet```  from Github with:
+
+```{r gh-installation, eval = FALSE}
 # install.packages("devtools")
 devtools::install_github("leoegidi/pivmet")
 ```
 
 ## Example 1. Dealing with label switching: relabelling in Bayesian mixture models by pivotal units (fish data)
 
-First of all, we load the package and we import the `fish` dataset
-belonging to the `bayesmix` package:
+First of all, we load the package and we import the ```fish``` dataset belonging to the ```bayesmix``` package:
 
-``` r
+```{r example}
+library(bayesmix)
 library(pivmet)
-#> Loading required package: bayesmix
-#> Loading required package: rjags
-#> Loading required package: coda
-#> Linked to JAGS 4.3.0
-#> Loaded modules: basemod,bugs
-#> Loading required package: mvtnorm
-#> Loading required package: RcmdrMisc
-#> Loading required package: car
-#> Loading required package: carData
-#> Loading required package: sandwich
-#> Warning: replacing previous import 'rstan::plot' by 'graphics::plot' when
-#> loading 'pivmet'
 data(fish)
 y <- fish[,1]
 N <- length(y)  # sample size 
@@ -52,47 +50,44 @@ k <- 5          # fixed number of clusters
 nMC <- 12000    # MCMC iterations
 ```
 
-Then we fit a Bayesian Gaussian mixture using the `piv_MCMC` function:
+Then we fit a Bayesian Gaussian mixture using the ```piv_MCMC``` function:
 
-``` r
+
+```{r fit, message =FALSE, warning = FALSE}
 res <- piv_MCMC(y = y, k = k, nMC = nMC)
-#> Compiling model graph
-#>    Declaring variables
[…]
