+++ man/piv_KMeans.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/musk.R
+% Please edit documentation in R/piv_KMeans.R
 \name{piv_KMeans}
 \alias{piv_KMeans}
 \title{k-means Clustering Using Pivotal Algorithms For Seeding}
