+++ man/piv_sim.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/sim.R
+% Please edit documentation in R/piv_sim.R
 \name{piv_sim}
 \alias{piv_sim}
 \title{Generate Data from a Gaussian Nested Mixture}
