+++ man/piv_rel.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/relab.R
+% Please edit documentation in R/piv_rel.R
 \name{piv_rel}
 \alias{piv_rel}
 \title{Performing the pivotal relabelling step and computing the relabelled posterior estimates}
