+++ man/piv_sel.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pivotal.R
+% Please edit documentation in R/piv_sel.R
 \name{piv_sel}
 \alias{piv_sel}
 \title{Pivotal Selection via Co-Association Matrix}
