+++ man/piv_MCMC.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bayes_mcmc.R
+% Please edit documentation in R/piv_MCMC.R
 \name{piv_MCMC}
 \alias{piv_MCMC}
 \title{JAGS/Stan Sampling for Gaussian Mixture Models and Clustering via Co-Association Matrix.}
