+++ man/piv_plot.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/plot.R
+% Please edit documentation in R/piv_plot.R
 \name{piv_plot}
 \alias{piv_plot}
 \title{Plotting outputs from pivotal relabelling}
