+++ DESCRIPTION
@@ -26,5 +26,5 @@ Imports: cluster, mclust, MASS,
           rjags, mvtnorm, bayesplot, scales
 Suggests: knitr, rmarkdown
 VignetteBuilder: knitr
-RoxygenNote: 7.2.1
+RoxygenNote: 7.3.1
 BuildManual: yes
