### 2024-05-09 e21d755
Create CODE OF CONDUCT.md
