+++ CODE OF CONDUCT.md
@@ -1,0 +1,38 @@
+# Contributor code of conduct
+
+
+As package mantainers and developers, we strongly invite any interested contributor/user to
+provide his/her experience in improving the package functionalities.  
+
+## How to contribute
+
+The goals of `pivmet` package are to:
+
+- construct a flexible environment for fitting Gaussian mixtures and sparse Gaussian mixtures;
+
+- undo the label switching problem;
+
+- perform a robust k-means clustering solution;
+
+- report graphical inspection of the final inferential conclusion.
+
+If you feel you can improve the areas above mentioned, or add some features, do not hesitate
+to contact us via the Github platform or by email (see below for further details about code and documentation).
+
+## Fix bugs
+
+Please, if you find a bug in the package, you can create a pull request or open an issue 
+in Github directly. 
+
+## Contribute documentation
+
+Documentation - vignettes, help pages - is a high priority for our package. If you feel to improve 
+documentation, please open a pull request or an issue in Github directly.
+
+## Contribute code
+
+If you are a more experienced R programmer, you may want to contribute directly to package development. Before you submit a pull request on our package, always file an issue and confirm the `pivmet` team agrees with your idea and is happy with your basic proposal.
+
+
+
+
