### 2024-05-09 LeoEgidi/pivmet#1 @LeoEgidi
Dear @adriancorrendo ,
I worked at the paper review. Thanks a lot for your valuable and precious comments.
I provide here a complete list of answers to your single points.
Of course, I am available for any clarification or doubt.

1. > I’m afraid there is no example in the paper of at least a line of code running any package functions. Could you please try to include something short with a line of code to run the key/s function/s?

Done. We added the relevant lines of R code for both the examples reported in the paper, see #Example 1 and #Example 2

2. > It sounds silly, but authors only indicate how to do it from github (although the code line to do it from CRAN is there). I think it could be good to use something like this on the README
:
** You can install the CRAN version of pivmet with:
[install.packages](https://rdrr.io/r/utils/install.packages.html)("pivmet")

** You can install the development version from [GitHub](https://github.com/) with: …

Thanks, done: we added the suggested intructions in the README.rmd and README.md files.

3. > Pivmet repo does not include any automated tests through continuous integration tools. @LeoEgidi, could you please revise this point? I could not find these tests like: R-CMD, codecov, appveyor.

Done. We added new tests in the 'tests/testthat' subfolder to monitor and check the performance of the main functions of the package.

4.  > I could not find community guidelines documentation. Please, see an example here…https://www.tidymodels.org/contribute/

Thanks, done. We included a new "CODE OF CONDUCT" in the main repo, where we declare and suggest how to contribute and report bugs.

5. > I consider authors fail to provide a comparison of what potentially similar packages are available or mentioning how do people is figuring out for similar tasks without pivmet. Could you please offer a brief overview to readers?

Done. We added a new paragraph in the paper, at the end of the "Statement of need", where we offer a brief overview of the main packages related to pivmet.

6. > Please, could you check for a potential typo at line 59 where it says..."Figure ??"

Done, it should be fixed right now.
