### 2024-05-09 #6461 @LeoEgidi
> > Hi @adriancorrendo , thanks a lot for your comments, I took a first look. I'll try to get a deeper look this week, or at most next week. Should I reply your points here or directly in the repo? What do you prefer? Thanks a lot
>
> Hi Leo, how are you? You would need to answer on the issue I created in the package repo. Once you address those comments, I'll start checking the remaining boxes until giving my full ok. That's what I did when I published my paper here.
>
> Hope it helps! Adrian

@arfon @adriancorrendo

I replied the comments from Adrian in the issue he created. Now time to work on the @larryshamalama comments
