### 2024-05-13 LeoEgidi/pivmet#1 @adriancorrendo
Thank you, @LeoEgidi ! Great job with the reviews.
Just have a comment regarding #3 Automated tests.
Including the tests with testthat is fine, but I meant adding something like R-CMD, codecov, "passing" badges on the readme. With these badges people will instantly see if the package passed the tests. Ideally, codecov will show what percentage of functions are "covered" with the tests you specified. Could you try at least adding the R-CMD ?

To add the R-CMD for example, you can use usethis::use_github_action()

Thanks,
Adrian
