### 2024-05-15 #6461 @adriancorrendo
I just asked authors to add a R-CMD badge in the README that helps users to see upfront if the package is passing automated tests, ideally also a coverage test like codecov. However, this could be more a "suggestion" than a required revision for acceptance. So, I'm happy with the answers and revisions from the authors.
