### 2024-05-15 LeoEgidi/pivmet#1 @adriancorrendo
Hi @LeoEgidi , please take my last comment as a "suggestion" rather than a required revision. I'm happy with your work on my revisions.

Congratulations for your package.

Best regards,
Adrian
