### 2024-05-16 0d10762
add r-cmd in readme
