### 2024-05-16 LeoEgidi/pivmet#1 @LeoEgidi
Thanks a lot @adriancorrendo for your nice suggestions and approval!
I added the R-CMD in the readme through your suggested function, thanks a lot!
