### 2024-05-16 #6461 @LeoEgidi
Thanks a lot!!

I also added the R-CMD suggestion from @adriancorrendo  in the readme
