+++ paper/paper.md
@@ -36,15 +36,15 @@ affiliations:
 # Summary
   
 We introduce the `R` package `pivmet`, a software that performs different pivotal methods for identifying, extracting, and using 
-the so-called pivotal units of a dataset that are chosen to represent the groups of data points to which they belong. 
-These algorithms turn out be very useful in many unsupervised and supervised learning frameworks such as clustering, 
+the so-called pivotal units  that are chosen from a partition of data points to represent the groups to which they belong. 
+Such units turn out be very useful in both unsupervised and supervised learning frameworks such as clustering, 
 classification and mixture modelling.  
-More specifically, applications of pivotal methods could cover, among the others: a Markov-Chain Monte Carlo (MCMC) 
-relabelling procedure to deal with the well-known label-switching problem [@stephens2000dealing; @richardson1997bayesian; 
-@fruhwirth2001markov; @egidi2018relabelling]  occurring during Bayesian estimation of mixture models; 
+More specifically, applications of pivotal methods include, among the others: a Markov-Chain Monte Carlo (MCMC) 
+relabelling procedure to deal with the well-known label-switching problem occurring during Bayesian estimation of mixture models [@stephens2000dealing; @richardson1997bayesian; 
+@fruhwirth2001markov; @egidi2018relabelling]; 
 model-based clustering through sparse finite mixture models (SFMM) [@malsiner2016model; @fruhwirth2019here]; consensus clustering [@JMLR02],
 which may allow to improve classical clustering techniques---e.g. the classical $k$-means---via a careful seeding; 
-and Dirichlet process mixture models (DPMM) [@ferguson1973bayesian; @escobar1995bayesian; @neal2000markov] in Bayesian nonparametrics. 
+and Dirichlet process mixture models (DPMM) in Bayesian nonparametrics [@ferguson1973bayesian; @escobar1995bayesian; @neal2000markov]. 
 
 ## Installation
 
@@ -60,22 +60,23 @@ However, before installing the package, the user should  make sure to download t
 
 # Statement of need
 
-In the modern *big-data* and *machine learning* age, summarizing some essential information from a data pattern is often relevant and can
+In the modern *big-data* and *machine learning* age, summarizing some essential information from a dataset is often relevant and can
 help simplifying the data pre-processing steps. The advantage of identifying representative units of a group---hereafter *pivotal units* 
-or *pivots*---somehow chosen to be  as far as possible from units in the other groups and as similar as possible to the units in the same 
-group is that they may convey relevant information about the group they belong to while saving wasteful operations.  
+or *pivots*---chosen in such a way that they are as far as possible from units in the other groups and/or as similar as possible to the units in the same 
+group, is that they may convey relevant information about the group they belong to while saving wasteful operations.  
 Despite the lack of a strict theoretical framework behind their characterization, the pivots may be beneficial in many machine learning frameworks,
-such as clustering, classification, and mixture modelling to derive reliable estimates and/or a better grouping partition.  
-A deep and theoretical detail around the package's supported pivotal methods is provided in [@egidi2018relabelling].
+such as clustering, classification, and mixture modelling when the interest is in deriving reliable estimates in mixture models and/or finding a partition of the data points.  
+The theoretical framework concerning the pivotal methods implemented in the `pivmet` package is provided in [@egidi2018relabelling].
 
-The `pivmet`  package [@pivmet] for `R` , available from the Comprehensive ` R`  Archive Network (CRAN) at
-[http://CRAN.R-project.org/package=pivmet](http://CRAN.R-project.org/package=pivmet), implements various pivotal selection criteria to 
[…]
