+++ paper/paper.bib
@@ -138,7 +138,7 @@ pages={583-617}
 
 
 @article{fruhwirth2001markov,
-  title={Markov chain Monte Carlo estimation of classical and dynamic switching and mixture models},
+  title={Markov {C}hain {M}onte {C}arlo estimation of classical and dynamic switching and mixture models},
   author={Fr{\"u}hwirth-Schnatter, Sylvia},
   journal={Journal of the American Statistical Association},
   volume={96},
@@ -301,7 +301,7 @@ pages = {313--331},
 }
 
 @article{richardson1997bayesian,
-  title={On Bayesian analysis of mixtures with an unknown number of components (with discussion)},
+  title={On {B}ayesian analysis of mixtures with an unknown number of components (with discussion)},
   author={Richardson, Sylvia and Green, Peter J},
   journal={Journal of the Royal Statistical Society: series B},
   volume={59},
@@ -417,7 +417,7 @@ Classification, and Density Estimation},
 }
 
 @article{neal2000markov,
-  title={Markov chain sampling methods for Dirichlet process mixture models},
+  title={Markov {C}hain sampling methods for {D}irichlet process mixture models},
   author={Neal, Radford M},
   journal={Journal of computational and graphical statistics},
   volume={9},
@@ -465,7 +465,7 @@ Classification, and Density Estimation},
 }
 
 @article{fruhwirth2019here,
-  title={From here to infinity: sparse finite versus Dirichlet process mixtures in model-based clustering},
+  title={From here to infinity: sparse finite versus {D}irichlet process mixtures in model-based clustering},
   author={Fr{\"u}hwirth-Schnatter, Sylvia and Malsiner-Walli, Gertraud},
   journal={Advances in data analysis and classification},
   volume={13},
@@ -477,7 +477,7 @@ Classification, and Density Estimation},
 
 
 @article{malsiner2016model,
-  title={Model-based clustering based on sparse finite Gaussian mixtures},
+  title={Model-based clustering based on sparse finite {G}aussian mixtures},
   author={Malsiner-Walli, Gertraud and Fr{\"u}hwirth-Schnatter, Sylvia and Gr{\"u}n, Bettina},
   journal={Statistics and computing},
   volume={26},
