### 2024-05-29 #6461 @LeoEgidi
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * Double check authors and affiliations (including ORCIDs)

     DONE

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

     DONE, the current version for pivmet is 0.6.0 (available on Github and submitted to CRAN)

> * Archive the release on Zenodo/figshare/etc and post the DOI here.

     DONE. DOI is: 10.5281/zenodo.11243277

> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

     DONE

> * Make sure that the license listed for the archive is the same as the software license.

     DONE (License is GPL-2)

Here we are @skanwal
