### 2024-05-30 #6461 @LeoEgidi
> @LeoEgidi - did you use [similar instructions](https://open.win.ox.ac.uk/pages/open-science/community/Open-WIN-Community/docs/gitlab/4-2-you-doi/#:~:text=In%20zenodo%2C%20log%20in%20or,on%20GitLab%2C%20or%20other%20documentation.) as this to create DOI?

yes!
I copy here the whole DOI url, this works:

https://doi.org/10.5281/zenodo.11243277
