### 2024-06-11 #6461 @LeoEgidi
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/9456754322?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ```yaml
> Element italic is not declared in institution list of possible children
> ```

Hi, I guess this is related with Italic fonts in the paper. I tried to fix, sorry. Let me know if you have further clues and if everything fixes when recommending acceptance
