### 2024-06-11 #6461 @LeoEgidi
> 👋 @openjournals/dsais-eics, this paper is ready to be accepted and published.
>
> Check final proof 👉📄 [Download article](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06461/joss.06461/10.21105.joss.06461.pdf)
>
> If the paper PDF and the deposit XML files look good in [openjournals/joss-papers#5477](https://github.com/openjournals/joss-papers/pull/5477), then you can now move forward with accepting the submission by compiling again with the command `@editorialbot accept`

Proofs are ok for me, thanks
