### 2024-06-11 #6461 @crvernon
:wave: @LeoEgidi  (cc. @skanwal)

There are a few issues that need to be addressed before this one is ready for publication:

- [ ] The license being declared in the R package via the description file is `GPL-2`.  This also seems like what was being referred to in the `LICENSE.txt` file - though this does not seem to represent the license in full and verbatim (see https://opensource.org/license/gpl-2-0).  This is also not the license that is declared in the companion Zenodo archive (you state v3, see https://zenodo.org/records/11243277).  Please make sure these are consistent and represented correctly.
- [ ] Your header on your readme is not working and your badge is showing failing tests.  Please clean these up.

- [ ] The header level for "Installation" should be the same as the summary or statement of need section.  In fact, you can remove the installation section as these instructions are present in your documentation.
- [ ] LINE 36:  Missing a new line before the paragraph start.
- [ ] LINE 40: Correct the citation format you stated as "package is provided in (Egidi et al., 2018)." This should be stated as "package is provided in Egidi et al. (2018)."
- [ ] Figure 2 should appear after it is mentioned in the text.
- [ ] LINE 109:  Is something missing in the `(,,)` section of "Figure 2 displays the traceplots for the parameters (, , )."?
- [ ] Please go through all references and add in any missing DOI where available. The check references output provided above in this thread provides many recommendations here to check.
- [ ] LINE 167:  "bayesian" should have a capital "B".  You can maintain capitalization in your references by wrapping any text that needs to maintain its formatting in curly brackets.

Please let me know once these have been addressed.  Thank you!
