+++ paper/paper.bib
@@ -24,7 +24,8 @@ author={Egidi, Leonardo and Pappad{\`a}, Roberta and Pauli, Francesco and Torell
   year = 2018,
   editor={C. Perna and M. Pratesi and A. Ruiz-Gazen} ,
   publisher={Springer}, 
- pages={71--81}
+ pages={71--81},
+ doi = {https://doi.org/10.1007/978-3-319-73906-9_7}
  }
 
 
@@ -37,7 +38,8 @@ author={Egidi, Leonardo and Pappad{\`a}, Roberta and Pauli, Francesco and Torell
   number={4},
   pages={957--969},
   year={2018},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {https://doi.org/10.1007/s11222-017-9774-2}
 }
 
 
@@ -48,7 +50,8 @@ title={Cluster ensembles - A knowledge reuse framework for combining multiple pa
 journal={Journal on Machine Learning Research},
 year={2002},
 volume={3},
-pages={583-617}
+pages={583-617},
+doi = {https://doi.org/10.1162/153244303321897735}
 }
 
 
@@ -60,7 +63,8 @@ pages={583-617}
  volume = {27},
  number = {6},
  year = {2005},
- pages = {835--850}
+ pages = {835--850},
+ doi = {10.1109/TPAMI.2005.113}
 } 
 
 
@@ -79,7 +83,8 @@ pages={583-617}
   booktitle={Advances in Intelligent Data Analysis VIII},
   pages={381--392},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {https://doi.org/10.1007/978-3-642-03915-7_33}
 }
 
 @article{fritsch2009improved,
@@ -90,7 +95,8 @@ pages={583-617}
   number={2},
   pages={367--391},
   year={2009},
-  publisher={International Society for Bayesian Analysis}
+  publisher={International Society for Bayesian Analysis},
+  doi = {10.1214/09-BA414}
 }
 
 @article{stephens2000dealing,
@@ -101,7 +107,8 @@ pages={583-617}
   number={4},
   pages={795--809},
   year={2000},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {https://doi.org/10.1111/1467-9868.00265}
 }
 
 @article{yao2012bayesian,
