+++ paper/paper.bib
@@ -78,7 +78,7 @@ doi = {https://doi.org/10.1162/153244303321897735}
 
 
 @incollection{puolamaki2009bayesian,
-  title={Bayesian solutions to the label switching problem},
+  title={{B}ayesian solutions to the label switching problem},
   author={Puolam{\"a}ki, Kai and Kaski, Samuel},
   booktitle={Advances in Intelligent Data Analysis VIII},
   pages={381--392},
@@ -90,12 +90,12 @@ doi = {https://doi.org/10.1162/153244303321897735}
 @article{fritsch2009improved,
   title={Improved criteria for clustering based on the posterior similarity matrix},
   author={Fritsch, Arno and Ickstadt, Katja},
-  journal={Bayesian analysis},
+  journal={{B}ayesian analysis},
   volume={4},
   number={2},
   pages={367--391},
   year={2009},
-  publisher={International Society for Bayesian Analysis},
+  publisher={International Society for {B}ayesian Analysis},
   doi = {10.1214/09-BA414}
 }
 
@@ -112,7 +112,7 @@ doi = {https://doi.org/10.1162/153244303321897735}
 }
 
 @article{yao2012bayesian,
-  title={Bayesian mixture labeling and clustering},
+  title={{B}ayesian mixture labeling and clustering},
   author={Yao, Weixin},
   journal={Communications in Statistics-Theory and Methods},
   volume={41},
@@ -135,7 +135,7 @@ doi = {https://doi.org/10.1162/153244303321897735}
 
 
 @article{yao2009bayesian,
-  title={Bayesian mixture labeling by highest posterior density},
+  title={{B}ayesian mixture labeling by highest posterior density},
   author={Yao, Weixin and Lindsay, Bruce G},
   journal={Journal of the American Statistical Association},
   volume={104},
@@ -152,7 +152,8 @@ doi = {https://doi.org/10.1162/153244303321897735}
   number={453},
   pages={194--209},
   year={2001},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi ={https://doi.org/10.1198/016214501750333063}
 }
 
 @article{chung2004difficulties,
@@ -194,7 +195,8 @@ pages = {313--331},
 	year = {2016},
 	volume = {69},
 	number = {1},
-	pages = {1--24}
+	pages = {1--24},
+	doi = {10.18637/jss.v069.c01}
 }
 
 @Manual{label.switching,
@@ -258,7 +260,7 @@ pages = {313--331},
 }
 
 @article{marin2005bayesian,
-  title={Bayesian modelling and inference on mixtures of distributions},
+  title={{B}ayesian modelling and inference on mixtures of distributions},
   author={Marin, Jean-Michel and Mengersen, Kerrie and Robert, Christian P},
   journal={Handbook of {S}tatistics},
   volume={25},
@@ -290,7 +292,7 @@ pages = {313--331},
 }
 
 @phdthesis{jasra2006bayesian,
-  title={Bayesian inference for mixture models via {M}onte {C}arlo computation},
+  title={{B}ayesian inference for mixture models via {M}onte {C}arlo computation},
   author={Jasra, Ajay},
   year={2006},
   school={Imperial College London (University of London)}
@@ -315,7 +317,8 @@ pages = {313--331},
   number={4},
   pages={731--792},
   year={1997},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi ={https://doi.org/10.1111/1467-9868.00095}
 }
 
 @article{roeder1990density,
@@ -356,10 +359,10 @@ Classification, and Density Estimation},
 }
 
 @Manual{pivmet,
[…]
