+++ paper/paper.md
@@ -66,6 +66,7 @@ In the modern *big-data* and *machine learning* age, summarizing some essential 
 help simplifying the data pre-processing steps. The advantage of identifying representative units of a group---hereafter *pivotal units* 
 or *pivots*---chosen in such a way that they are as far as possible from units in the other groups and/or as similar as possible to the units in the same 
 group, is that they may convey relevant information about the group they belong to while saving wasteful operations.  
+
 Despite the lack of a strict theoretical framework behind their characterization, the pivots may be beneficial in many machine learning frameworks,
 such as clustering, classification, and mixture modelling when the interest is in deriving reliable estimates in mixture models and/or finding a partition of the data points.  The theoretical framework concerning the pivotal methods implemented in the `pivmet` package is provided in @egidi2018relabelling.
 
