+++ paper/paper.bib
@@ -25,7 +25,7 @@ author={Egidi, Leonardo and Pappad{\`a}, Roberta and Pauli, Francesco and Torell
   editor={C. Perna and M. Pratesi and A. Ruiz-Gazen} ,
   publisher={Springer}, 
  pages={71--81},
- doi = {https://doi.org/10.1007/978-3-319-73906-9_7}
+ doi = {10.1007/978-3-319-73906-9_7}
  }
 
 
@@ -39,7 +39,7 @@ author={Egidi, Leonardo and Pappad{\`a}, Roberta and Pauli, Francesco and Torell
   pages={957--969},
   year={2018},
   publisher={Springer},
-  doi = {https://doi.org/10.1007/s11222-017-9774-2}
+  doi = {10.1007/s11222-017-9774-2}
 }
 
 
@@ -51,7 +51,7 @@ journal={Journal on Machine Learning Research},
 year={2002},
 volume={3},
 pages={583-617},
-doi = {https://doi.org/10.1162/153244303321897735}
+doi = {10.1162/153244303321897735}
 }
 
 
@@ -84,7 +84,7 @@ doi = {https://doi.org/10.1162/153244303321897735}
   pages={381--392},
   year={2009},
   publisher={Springer},
-  doi = {https://doi.org/10.1007/978-3-642-03915-7_33}
+  doi = {10.1007/978-3-642-03915-7_33}
 }
 
 @article{fritsch2009improved,
@@ -108,7 +108,7 @@ doi = {https://doi.org/10.1162/153244303321897735}
   pages={795--809},
   year={2000},
   publisher={Wiley Online Library},
-  doi = {https://doi.org/10.1111/1467-9868.00265}
+  doi = {10.1111/1467-9868.00265}
 }
 
 @article{yao2012bayesian,
@@ -153,7 +153,7 @@ doi = {https://doi.org/10.1162/153244303321897735}
   pages={194--209},
   year={2001},
   publisher={Taylor \& Francis},
-  doi ={https://doi.org/10.1198/016214501750333063}
+  doi ={10.1198/016214501750333063}
 }
 
 @article{chung2004difficulties,
@@ -318,7 +318,7 @@ pages = {313--331},
   pages={731--792},
   year={1997},
   publisher={Wiley Online Library},
-  doi ={https://doi.org/10.1111/1467-9868.00095}
+  doi ={10.1111/1467-9868.00095}
 }
 
 @article{roeder1990density,
@@ -435,7 +435,7 @@ Classification, and Density Estimation},
   pages={249--265},
   year={2000},
   publisher={Taylor \& Francis},
-  doi ={https://doi.org/10.1080/10618600.2000.10474879}
+  doi ={10.1080/10618600.2000.10474879}
 }
 
 @incollection{ferguson1983bayesian,
@@ -456,7 +456,7 @@ Classification, and Density Estimation},
   pages={577--588},
   year={1995},
   publisher={Taylor \& Francis},
-  doi = {https://doi.org/10.1080/01621459.1995.10476550}
+  doi = {10.1080/01621459.1995.10476550}
 }
 
 @misc{ross2018dirichletprocess,
@@ -485,7 +485,7 @@ Classification, and Density Estimation},
   pages={33--64},
   year={2019},
   publisher={Springer},
-  doi = {https://doi.org/10.1007/s11634-018-0329-y}
+  doi = {10.1007/s11634-018-0329-y}
 }
 
 
@@ -498,7 +498,7 @@ Classification, and Density Estimation},
   pages={303--324},
   year={2016},
   publisher={Springer},
-  doi = {https://doi.org/10.1007/s11222-014-9500-2}
+  doi = {10.1007/s11222-014-9500-2}
 }
 
 @article{greve2020spying,
