### 2024-06-12 #6461 @LeoEgidi
@skanwal @crvernon Thanks a lot!
I managed to fix your last points, here below is my review.

> 👋 @LeoEgidi (cc. @skanwal)
>
> There are a few issues that need to be addressed before this one is ready for publication:
>
> * [x]  The license being declared in the R package via the description file is `GPL-2`.  This also seems like what was being referred to in the `LICENSE.txt` file - though this does not seem to represent the license in full and verbatim (see https://opensource.org/license/gpl-2-0).  This is also not the license that is declared in the companion Zenodo archive (you state v3, see https://zenodo.org/records/11243277).  Please make sure these are consistent and represented correctly.

DONE. The License file has been correctly updated, you can check it on the Github page as well.
Moreover, the License in Zenodo is now correctly specified as GPL 2.0 as well.

> * [ ]  Your header on your readme is not working and your badge is showing failing tests.  Please clean these up.

DONE. The header now has been suppressed.

Regarding your second point, I deeply checked the failing tests, and I realized there is a 'false positive' here.
Precisely, the problem with these checks arises because windows-latest (release) and macos-latest (release) can not
install the JAGS compiler (that is needed to launch the rjags package), whereas Ubuntu is successfull.
However, this is a common problem for many other packages using JAGS as an underlying compiler.
I deeply checked the package with automated checks and no errors or warnings arise.
If you take a look here, in fact, no errors arise:

https://cran.r-project.org/web/checks/check_results_pivmet.html

Moreover, the README clearly states that the user needs to install the JAGS compiler before installing the 'pivmet' package.
Anyway, for completeness, I copy and paste here below the JAGS error as given by Windows release:

'Make sure you have installed JAGS-4.x.y.exe (for any x >=0, y>=0) from
http://www.sourceforge.net/projects/mcmc-jags/files'

> In your paper:
>
> * [ ]  The header level for "Installation" should be the same as the summary or statement of need section.  In fact, you can remove the installation section as these instructions are present in your documentation.

DONE, same level now for the "Installation" section.

> * [ ]  LINE 36:  Missing a new line before the paragraph start.

DONE

> * [ ]  LINE 40: Correct the citation format you stated as "package is provided in (Egidi et al., 2018)." This should be stated as "package is provided in Egidi et al. (2018)."

DONE

> * [ ]  Figure 2 should appear after it is mentioned in the text.

DONE

> * [ ]  LINE 109:  Is something missing in the `(,,)` section of "Figure 2 displays the traceplots for the parameters (, , )."?

CORRECTED

> * [ ]  Please go through all references and add in any missing DOI where available. The check references output provided above in this thread provides many recommendations here to check.

DONE

> * [ ]  LINE 167:  "bayesian" should have a capital "B".  You can maintain capitalization in your references by wrapping any text that needs to maintain its formatting in curly brackets.

DONE

> Please let me know once these have been addressed. Thank you!

One last issue: due to the image sizes, I've replaced the old .png Figure 1 with a better pdf version.
