+++ paper/paper.bib
@@ -1,34 +1,3 @@
-@article{betancourt2017conceptual,
-  title={A conceptual introduction to Hamiltonian Monte Carlo},
-  author={Betancourt, Michael},
-  journal={arXiv preprint arXiv:1701.02434},
-  year={2017}
-}
-
-
-@incollection{kmeans,
-  author={Egidi, Leonardo and Pappad{\`a}, Roberta and Pauli, Francesco and Torelli, Nicola},
-  title = {K-means seeding via {M}{U}{S} algorithm},
-  booktitle = {Book of Short Papers SIS 2018},
-  year = 2018,
-  editor={A. Abbruzzo and E. Brentari and M. Chiodi and D. Piacentino},
-  publisher ={Pearson}, 
-  pages={256--262}
- }
-
-
-@incollection{egidi2018maxima,
-author={Egidi, Leonardo and Pappad{\`a}, Roberta and Pauli, Francesco and Torelli, Nicola},
-  title = {Maxima {U}nits {S}earch ({M}{U}{S}) algorithm: methodology and applications},
-  booktitle = {Studies in Theoretical and Applied Statistics},
-  year = 2018,
-  editor={C. Perna and M. Pratesi and A. Ruiz-Gazen} ,
-  publisher={Springer}, 
- pages={71--81},
- doi = {10.1007/978-3-319-73906-9_7}
- }
-
-
 
 @article{egidi2018relabelling,
   title={Relabelling in {B}ayesian mixture models by pivotal units},
@@ -45,59 +14,16 @@ author={Egidi, Leonardo and Pappad{\`a}, Roberta and Pauli, Francesco and Torell
 
 
 @ARTICLE{JMLR02,
-author={Strehl, A. and Ghosh, J.},
-title={Cluster ensembles - A knowledge reuse framework for combining multiple partitions},
-journal={Journal on Machine Learning Research},
-year={2002},
-volume={3},
-pages={583-617},
-doi = {10.1162/153244303321897735}
+  author={Strehl, A. and Ghosh, J.},
+   title={Cluster ensembles - A knowledge reuse framework for combining multiple partitions},
+   journal={Journal on Machine Learning Research},
+   year={2002},
+   volume={3},
+   pages={583-617},
+   doi = {10.1162/153244303321897735}
 }
 
 
-
-@article{fredJain05,
- author = {Fred, Ana L. N. and Jain, Anil K.},
- title = {Combining multiple clusterings using evidence accumulation},
- journal = {IEEE Trans. Pattern Anal. Mach. Intell.},
- volume = {27},
- number = {6},
- year = {2005},
- pages = {835--850},
- doi = {10.1109/TPAMI.2005.113}
-} 
-
-
-@inproceedings{kmeans07,
-  title={k-means++: The advantages of careful seeding},
-  author={Arthur, David and Vassilvitskii, Sergei},
-  booktitle={Proceedings of the eighteenth annual ACM-SIAM symposium on Discrete algorithms},
-  pages={1027--1035},
-  year={2007}}
-
-
-
-@incollection{puolamaki2009bayesian,
-  title={{B}ayesian solutions to the label switching problem},
-  author={Puolam{\"a}ki, Kai and Kaski, Samuel},
-  booktitle={Advances in Intelligent Data Analysis VIII},
-  pages={381--392},
-  year={2009},
-  publisher={Springer},
-  doi = {10.1007/978-3-642-03915-7_33}
-}
-
-@article{fritsch2009improved,
-  title={Improved criteria for clustering based on the posterior similarity matrix},
-  author={Fritsch, Arno and Ickstadt, Katja},
-  journal={{B}ayesian analysis},
-  volume={4},
-  number={2},
[…]
