### 2024-06-12 #6461 @LeoEgidi
Thanks @crvernon

> @LeoEgidi thanks for making the changes. These last two changes should wrap things up:
>
> * [x]  LINE 13:  Missing new line before paragraph.

DONE

> * [ ]  Per the list of "Missing DOI" above, please remove any unused references from your paper.bib file

DONE
