### 2024-06-12 dc942d6
Create CITATION.cff
