### 2024-06-12 #6461 @crvernon
🥳 Congratulations on your new publication @LeoEgidi! Many thanks to @skanwal for editing and @adriancorrendo, and @larryshamalama for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
